# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Playbasis API",
        "description": "API Explorer",
        "termsOfService": "",
        "contact": {
            "email": "info@playbasis.com"
        },
        "version": "1.0"
    },
    "host": "api.pbapp.net",
    "basePath": "/",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/Trip": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "description": "List of Trips that the player has started.",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "trip_id",
                        "in": "query",
                        "description": "Trip ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "finished",
                        "in": "query",
                        "description": "Status of trip",
                        "required": false,
                        "type": "string",
                        "default": "",
                        "enum": [
                            "",
                            "true",
                            "false"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Trip/startTrip": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "description": "Start trip",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Trip/finishTrip": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "description": "Finish trip",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "drive_log",
                        "in": "query",
                        "description": "Response including driving log",
                        "required": true,
                        "type": "string",
                        "default": "false",
                        "enum": [
                            "true",
                            "false"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Trip/addTripLog": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "description": "Log driving status of the trip that the player has started",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token returned by Playbasis Authentication",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "trip_id",
                        "in": "query",
                        "description": "Trip Id",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "speed",
                        "in": "query",
                        "description": "Current speed | (km/h)",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Current latitude",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Current longitude",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "altitude",
                        "in": "query",
                        "description": "Current altitude",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "rpm",
                        "in": "query",
                        "description": "Current RPM",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "distance",
                        "in": "query",
                        "description": "Distance from start",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "speed_limit",
                        "in": "query",
                        "description": "Speed limit at the location | (km/h)",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/Trip/getTripLog": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "description": "Retrieve driving status log of specific trip Id",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "query",
                        "description": "Player ID as used in client's website",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "trip_id",
                        "in": "query",
                        "description": "Trip Id",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        },
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    },
    "security": [
        {
            "apiKey": [],
            "token": []
        }
    ]
}